feat: Add global spam count endpoint
This commit adds a new endpoint `/get_global_spam_count` to the webserver. This endpoint retrieves the global spam count from the `global_spam_count.json` file and returns it as a JSON response. The global spam count is used to keep track of the total number of spam emails sent.